Move log out logic from welcome to profile
